+++ org/xbill/DNS/DNAMERecord.java
@@ -20,7 +20,7 @@ private
 DNAMERecord() {}
 
 private
-DNAMERecord(Name name, int dclass, int ttl) {
+DNAMERecord(Name name, int dclass, long ttl) {
 	super(name, Type.DNAME, dclass, ttl);
 }
 
@@ -34,12 +34,12 @@ getMember() {
  * @param target The name to which the DNAME alias points
  */
 public
-DNAMERecord(Name name, int dclass, int ttl, Name target) {
+DNAMERecord(Name name, int dclass, long ttl, Name target) {
 	super(name, Type.DNAME, dclass, ttl, target);
 }
 
 Record
-rrFromWire(Name name, int type, int dclass, int ttl, int length,
+rrFromWire(Name name, int type, int dclass, long ttl, int length,
 	   DataByteInputStream in)
 throws IOException
 {
@@ -47,7 +47,7 @@ throws IOException
 }
 
 Record
-rdataFromString(Name name, int dclass, int ttl, Tokenizer st, Name origin)
+rdataFromString(Name name, int dclass, long ttl, Tokenizer st, Name origin)
 throws IOException
 {
 	return rdataFromString(new DNAMERecord(name, dclass, ttl), st, origin);
