+++ org/xbill/DNS/LOCRecord.java
@@ -24,7 +24,7 @@ private
 LOCRecord() {}
 
 private
-LOCRecord(Name name, int dclass, int ttl) {
+LOCRecord(Name name, int dclass, long ttl) {
 	super(name, Type.LOC, dclass, ttl);
 }
 
@@ -43,7 +43,7 @@ getMember() {
  * @param vPrecision The vertical precision of the data, in m.
 */
 public
-LOCRecord(Name name, int dclass, int ttl, double latitude, double longitude,
+LOCRecord(Name name, int dclass, long ttl, double latitude, double longitude,
 	  double altitude, double size, double hPrecision, double vPrecision)
 {
 	this(name, dclass, ttl);
@@ -56,7 +56,7 @@ LOCRecord(Name name, int dclass, int ttl, double latitude, double longitude,
 }
 
 Record
-rrFromWire(Name name, int type, int dclass, int ttl, int length,
+rrFromWire(Name name, int type, int dclass, long ttl, int length,
 	   DataByteInputStream in)
 throws IOException
 {
@@ -79,7 +79,7 @@ throws IOException
 }
 
 Record
-rdataFromString(Name name, int dclass, int ttl, Tokenizer st, Name origin)
+rdataFromString(Name name, int dclass, long ttl, Tokenizer st, Name origin)
 throws IOException
 {
 	LOCRecord rec = new LOCRecord(name, dclass, ttl);
