+++ org/xbill/DNS/NAPTRRecord.java
@@ -23,7 +23,7 @@ private Name replacement;
 private NAPTRRecord() {}
 
 private
-NAPTRRecord(Name name, int dclass, int ttl) {
+NAPTRRecord(Name name, int dclass, long ttl) {
 	super(name, Type.NAPTR, dclass, ttl);
 }
 
@@ -46,7 +46,7 @@ getMember() {
  * @throws IllegalArgumentException One of the strings has invalid escapes
  */
 public
-NAPTRRecord(Name name, int dclass, int ttl, int order, int preference,
+NAPTRRecord(Name name, int dclass, long ttl, int order, int preference,
 	    String flags, String service, String regexp, Name replacement)
 {
 	this(name, dclass, ttl);
@@ -66,7 +66,7 @@ NAPTRRecord(Name name, int dclass, int ttl, int order, int preference,
 }
 
 Record
-rrFromWire(Name name, int type, int dclass, int ttl, int length,
+rrFromWire(Name name, int type, int dclass, long ttl, int length,
 	   DataByteInputStream in)
 throws IOException
 {
@@ -83,7 +83,7 @@ throws IOException
 }
 
 Record
-rdataFromString(Name name, int dclass, int ttl, Tokenizer st, Name origin)
+rdataFromString(Name name, int dclass, long ttl, Tokenizer st, Name origin)
 throws IOException
 {
 	NAPTRRecord rec = new NAPTRRecord(name, dclass, ttl);
