+++ org/xbill/DNS/RPRecord.java
@@ -25,7 +25,7 @@ private
 RPRecord() {}
 
 private
-RPRecord(Name name, int dclass, int ttl) {
+RPRecord(Name name, int dclass, long ttl) {
 	super(name, Type.RP, dclass, ttl);
 }
 
@@ -40,7 +40,7 @@ getMember() {
  * @param textDomain The address where TXT records can be found
  */
 public
-RPRecord(Name name, int dclass, int ttl, Name mailbox, Name textDomain) {
+RPRecord(Name name, int dclass, long ttl, Name mailbox, Name textDomain) {
 	this(name, dclass, ttl);
 	if (!mailbox.isAbsolute())
 		throw new RelativeNameException(mailbox);
@@ -51,7 +51,7 @@ RPRecord(Name name, int dclass, int ttl, Name mailbox, Name textDomain) {
 }
 
 Record
-rrFromWire(Name name, int type, int dclass, int ttl, int length,
+rrFromWire(Name name, int type, int dclass, long ttl, int length,
 	   DataByteInputStream in)
 throws IOException
 {
@@ -64,7 +64,7 @@ throws IOException
 }
 
 Record
-rdataFromString(Name name, int dclass, int ttl, Tokenizer st, Name origin)
+rdataFromString(Name name, int dclass, long ttl, Tokenizer st, Name origin)
 throws IOException
 {
 	RPRecord rec = new RPRecord(name, dclass, ttl);
